#22 change return type of id() to int

+++ src/main/java/io/zold/api/RtTransaction.java
@@ -73,7 +73,7 @@ final class RtTransaction implements Transaction {
 
     @Override
     @SuppressWarnings("PMD.ShortMethodName")
-    public long id() throws IOException {
+    public int id() throws IOException {
         final String ident = new UncheckedText(
             new IoCheckedScalar<>(
                 new ItemAt<>(
